feat(proxy): Add comprehensive tests and utility functions for proxy setup
- Introduced a new test suite in `tests.rs` to validate proxy-related functionalities, including UID-based port assignment, proxy reachability, and environment variable exports for various shells.
- Added utility functions in `util.rs` for managing proxy configurations, including determining default ports based on user UID and checking proxy reachability.
- Implemented constants for proxy environment variable markers and notes for omitted configurations.
- Ensured compatibility with both Unix and non-Unix systems for UID-based port calculations.
