+++ rust/src/proxy/providers.rs
@@ -70,6 +70,16 @@ pub(super) fn stats_label<'a>(registry_id: Option<&str>, shape_label: &'a str) -
     }
 }
 
+/// True when an OpenAI-shaped registry request should use the Responses
+/// compressor (`input` / `function_call_output`) rather than Chat Completions.
+pub(super) fn is_openai_responses_path(path: &str) -> bool {
+    let path = path.trim_end_matches('/');
+    path == "/responses"
+        || path == "/v1/responses"
+        || path.starts_with("/responses/")
+        || path.starts_with("/v1/responses/")
+}
+
 pub async fn handler(
     State(state): State<ProxyState>,
     Path((id, rest)): Path<(String, String)>,
@@ -112,16 +122,35 @@ pub async fn handler(
             .await
         }
         WireShape::OpenAi => {
-            forward::forward_request(
-                State(state),
-                req,
-                &provider.base_url,
-                "/v1/chat/completions",
-                super::openai::compress_request_body,
-                "OpenAI",
-                &[],
-            )
-            .await
+            // Built-in OpenAI routes pick Chat Completions vs Responses by path
+            // (`/v1/chat/completions` vs `/v1/responses`). Registry providers
+            // must do the same: Grok CLI hits `/providers/grok-chat/v1/responses`
+            // with `function_call_output` in `input`, which the Chat compressor
+            // ignores (it only rewrites `messages`). Path already has the
+            // `/providers/{id}` prefix stripped above.
+            if is_openai_responses_path(req.uri().path()) {
+                forward::forward_request(
+                    State(state),
+                    req,
+                    &provider.base_url,
+                    "/v1/responses",
+                    super::openai_responses::compress_request_body,
+                    "OpenAI",
+                    &[],
+                )
+                .await
+            } else {
+                forward::forward_request(
+                    State(state),
+                    req,
+                    &provider.base_url,
+                    "/v1/chat/completions",
+                    super::openai::compress_request_body,
+                    "OpenAI",
+                    &[],
+                )
+                .await
+            }
         }
         WireShape::Gemini => {
             // Gemini carries the model in the URL path, not the body (#840).
@@ -219,6 +248,17 @@ mod tests {
         assert!(is_grok_provider_id(" Grok "));
         assert!(!is_grok_provider_id("openai"));
         assert!(!is_grok_provider_id("foundry"));
+    }
+
+    #[test]
+    fn is_openai_responses_path_detects_responses_api() {
+        assert!(is_openai_responses_path("/v1/responses"));
+        assert!(is_openai_responses_path("/v1/responses/"));
+        assert!(is_openai_responses_path("/responses"));
+        assert!(is_openai_responses_path("/v1/responses/resp_123/input_items"));
+        assert!(!is_openai_responses_path("/v1/chat/completions"));
+        assert!(!is_openai_responses_path("/v1/models"));
+        assert!(!is_openai_responses_path("/v1/responsesx"));
     }
 
     fn provider(shape: WireShape, api_key_env: Option<&str>) -> ResolvedProvider {
